Move ConfigLoader class to own file
